Rename "MuoVErsi" package to "server"
