+++ alembic/env.py
@@ -4,8 +4,8 @@ from alembic import context
 from sqlalchemy import engine_from_config
 from sqlalchemy import pool
 
-from MuoVErsi.base.models import Base
-from MuoVErsi.handlers import engine_url
+from server.base.models import Base
+from server.handlers import engine_url
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
