+++ run.py
@@ -1,6 +1,6 @@
 import asyncio
 
-from MuoVErsi.handlers import main
+from server.handlers import main
 
 if __name__ == "__main__":
     asyncio.run(main())
