+++ save_data.py
@@ -2,10 +2,10 @@ import logging
 
 from sqlalchemy.orm import sessionmaker
 
-from MuoVErsi.handlers import engine
-from MuoVErsi.trenitalia import Trenitalia
-from MuoVErsi.GTFS import GTFS
-from MuoVErsi.typesense import connect_to_typesense
+from server.GTFS import GTFS
+from server.handlers import engine
+from server.trenitalia import Trenitalia
+from server.typesense import connect_to_typesense
 
 logging.basicConfig(
     format="%(asctime)s - %(name)s - %(levelname)s - %(message)s", level=logging.INFO
