+++ server/trenitalia/source.py
@@ -1,12 +1,12 @@
 import json
 import math
 import os
-from pytz import timezone
 
 import requests
+from pytz import timezone
 from tqdm import tqdm
 
-from MuoVErsi.base import *
+from server.base import *
 
 logging.basicConfig(
     format="%(asctime)s - %(name)s - %(levelname)s - %(message)s", level=logging.INFO
